✨ feat: 답장 상세 조회 응답 400 필터링하는 ReceptionFallback 적용
