+++ src/components/ErrorBoundary/UnknownErrorBoundary.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { ErrorBoundary, type FallbackProps } from 'react-error-boundary';
-import { useLocation, useNavigate } from 'react-router-dom';
+import { useLocation } from 'react-router-dom';
 
 interface UnknownErrorBoundaryProps {
   FallbackComponent: React.ComponentType<FallbackProps>;
@@ -11,15 +11,10 @@ const UnknownErrorBoundary = ({
   FallbackComponent,
   children,
 }: UnknownErrorBoundaryProps) => {
-  const navigate = useNavigate();
   const { key } = useLocation();
 
   return (
-    <ErrorBoundary
-      FallbackComponent={FallbackComponent}
-      onReset={() => navigate(-1)}
-      resetKeys={[key]}
-    >
+    <ErrorBoundary FallbackComponent={FallbackComponent} resetKeys={[key]}>
       {children}
     </ErrorBoundary>
   );
